Fix - should not be running as root
This establishes a dedicated system account, named "mirrmaid", for the
purposes of running the process.

NOTE: This also relocates the config file!  This was done to establish a
suitable home-directory for the mirrmaid user, which will conveniently provide
a location for SSH keys if those happen to be needed for any mirroring
transport.