New - grouping/splitting of operations summaries
Via a new summary_group option in the mirrmaid configuration, it is now
possible to split/group the operations summaries when multiple jobs
exist on a host.  Previously, there was only one operations summary
regardless of how many mirrmaid jobs may have been established.  This
new feature allows the summaries to be grouped and/or split as desired.