Fix - some MTAs may require sender's domain
Depending on how the local MTA has been configured for mail submission,
a default domain may or may not be suffixed to the 'mirrmaid' sender
address.  If the MTA doesn't suffix a default domain, the mail
submission is likely to be rejected.  This revision thwarts that issue
by suffixing the fully qualified host name.