New - proxy support
Prior to this revision, it was only possible to use mirrmaid with
a proxy by exporting the RSYNC_PROXY environment variable prior to
starting the mirrmaid process.  This revision adds direct support, which
must be used instead of the RSYNC_PROXY method since that environment
variable will be cleared if the new "proxy" configuration setting has no
declared value.