New - notification summarizes reason
The header portion of the activity summary now includes a reason for the
notification having been sent.  This may help draw attention to the fact
that it is being sent due to size, which may tell of an immediate
failure or explain why messages are being sent more often than expected.
The more typical reason would be due to the age of the message, but
a reason of "forced" could also be seen under certain circumstances.