New - drop privileges at startup
If mirrmaid is started as any user other than mirrmaid (and group
mirrmaid), it will now attempt to drop privileges to that user/group and
proceed only if successful.  It has always been intended to be run as
this user/group and normally that is, in fact, the case.  However, it
sometimes gets accidentally run as root/root, perhaps through sudo and
haste, by some user attempting to run it manually, which may botch
ownership of newly created files and consequently prevent those files
from being updated later when the correct user/group is indeed used.