Fix - ConnectionError exception is unhandled
When the LogSummarizingHandler attempts to deliver the summary via
email, it's possible for a ConnectionError to be raised which results in
mirrmaid fatality.  Due to the design this could easily result in the
termination occurring prior to any mirror synchronization, thus it's
imperative that the ConnectionError be caught and dealt with.  In this
case, the exception merely results in a message written to stderr.  It
would be good to log the message as well, but adding anything more to
the log within this handler simply results in a recursive call to
trim/rotate and summarize the log again; i.e., an infinite loop that
quickly exhausts the stack.