New - "koji" and "koji-dart" release targets
This establishes new release targets for tito which will inject
a buildSRPMFromSCM task into Koji.  That task expects a make target
named "sources" which is to produce/obtain a typical upstream tarball of
the sources.  This commit provides a Makefile that will dynamically
create such an archive on the fly.  This archive will be named and
contain sources matching the Git tag being built, just as Koji will
expect.