Change - use soft static user/group account allocation
This commit makes this package better conform to the Fedora Packaging
Guidelines with respect to the allocation strategy for the "mirrmaid"
system user/group account.  The dynamic allocation method previously
used has been replaced here with FPC's scriptlet for soft static
allocation, although the UID/GID used here are not officially assigned
by FPC so there's potential for conflict in the future.