Bug - FileNotFoundError raised in log rollover
http://bugs.python.org/issue18940 was resolved poorly by introducing
a race condition, but perhaps the RotatingFileHandler wasn't intended
for multi-process access to the same log target.  If the source is
missing, we can assume that another process has already performed the
rollover.

This might indicate a need for logs to be separated by mirror (and
consequently protected by the lock of each), but there are no other
known problems with concurrent access to the logs in mirrmaid, so far
now this commit, though a bit of a kludge, seems like a reasonable
compromise.