Bug - FileNotFoundError raised in log rollover
This is a followup to commit ef6100a which unfortunately did not resolve
the problem.  Further analysis has revealed that the problem was either
never within our LogSummarizingHandler's rotator or that was not the
only case.  For certain, the FileNotFoundError exception can be raised
by the BaseRotatingHandler's rotate() method because it clearly has
a race condition via:

    if os.path.exists(source):
        os.rename(source, dest)

Honestly, how does crap like that get into the standard library?